requirements.txt
futures
futurist
httplib2
ipaddress
paramiko
PyYAML
requests
requests-futures
sh
urllib3
monotonic
